serving/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
redis==4.6.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-otlp==1.20.0
prometheus-client==0.17.1

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0
